+++ src/gash/router/server/queue/GlobalCommandInboundAppWorker.java
@@ -13,34 +13,29 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package gash.router.server.queue.work;
+package gash.router.server.queue;
 
 import com.google.protobuf.GeneratedMessage;
-import gash.router.server.MessageServer;
-import gash.router.server.PrintUtil;
-import gash.router.server.edges.EdgeInfo;
-import gash.router.server.resources.Failure;
 import gash.router.server.resources.Ping;
-import gash.router.server.resources.Task;
+import gash.router.server.resources.Query;
+import global.Global;
 import io.netty.channel.Channel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import pipe.common.Common;
-import pipe.work.Work;
 
-public class WorkInboundAppWorker extends Thread {
+public class GlobalCommandInboundAppWorker extends Thread {
 	protected static Logger logger = LoggerFactory.getLogger("server");
 
 	int workerId;
-	PerChannelWorkQueue sq;
+	PerChannelGlobalCommandQueue sq;
 	boolean forever = true;
 
-	public WorkInboundAppWorker(ThreadGroup tgrp, int workerId, PerChannelWorkQueue sq) {
+	public GlobalCommandInboundAppWorker(ThreadGroup tgrp, int workerId, PerChannelGlobalCommandQueue sq) {
 		super(tgrp, "inboundWork-" + workerId);
 		this.workerId = workerId;
 		this.sq = sq;
 
-		if (sq.inbound == null)
+		if (sq.inboundWork == null)
 			throw new RuntimeException("connection worker detected null inboundWork queue");
 	}
 
@@ -53,34 +48,32 @@ public class WorkInboundAppWorker extends Thread {
 		}
 
 		while (true) {
-			if (!forever && sq.inbound.size() == 0)
+			if (!forever && sq.inboundWork.size() == 0)
 				break;
 
 			try {
 				// block until a message is enqueued
-				GeneratedMessage msg = sq.inbound.take();
+				GeneratedMessage msg = sq.inboundWork.take();
 
 				// process request and enqueue response
+				if(msg instanceof Global.GlobalCommandMessage){
 
-				if (msg instanceof Work.WorkRequest) {
-					Work.WorkRequest req = ((Work.WorkRequest) msg);
-					Work.Payload payload = req.getPayload();
+					//PrintUtil.printCommand((Pipe.CommandRequest) msg);
+					Global.GlobalCommandMessage req = ((Global.GlobalCommandMessage) msg);
 
-					//PrintUtil.printWork(req);
-					if (payload.hasBeat()) {
-						//Work.Heartbeat hb = payload.getBeat();
-						logger.info("heartbeat from " + req.getHeader().getNodeId());
-					} else if (payload.hasPing()) {
+					if (req.hasPing()) {
 						new Ping(sq).handle(req);
+					} else if (req.hasMessage()) {
+						logger.info(req.getMessage());
+					} else if (req.hasQuery()){
+						new Query(sq).handle(req);
+					} else if(req.hasResponse()){
 
-					} else if (payload.hasErr()) {
-						new Failure().handle(req);
-						// PrintUtil.printFailure(err);
-					} else if (payload.hasTask()) {
-						new Task(sq).handle(req);
-					} else if (payload.hasState()) {
-						Work.WorkState s = payload.getState();
 					}
+					else{
+						//todo
+					}
+
 				}
 			} catch (InterruptedException ie) {
 				break;
@@ -91,7 +84,7 @@ public class WorkInboundAppWorker extends Thread {
 		}
 
 		if (!forever) {
-			logger.info("Work incoming connection queue closing");
+			logger.info("Command incoming connection queue closing");
 		}
 	}
 }
