+++ src/gash/router/server/resources/Failure.java
@@ -1,6 +1,7 @@
 package gash.router.server.resources;
 
 import com.google.protobuf.GeneratedMessage;
+import gash.router.server.queue.ChannelQueue;
 import global.Global;
 import pipe.common.Common;
 import pipe.work.Work;
@@ -11,8 +12,8 @@ import routing.Pipe;
  */
 public class Failure extends Resource {
 
-    public Failure(){
-
+    public Failure(ChannelQueue sq){
+        super(sq);
     }
 
     public void handleGlobalCommand(Global.GlobalCommandMessage msg) {
