+++ src/gash/router/server/resources/Query.java
@@ -1,10 +1,15 @@
 package gash.router.server.resources;
 
+import database.dao.MongoDAO;
+import database.model.MongoDataModel;
+import gash.router.server.queue.ChannelQueue;
+import gash.router.server.queue.PerChannelGlobalCommandQueue;
 import global.Global;
-import pipe.common.Common;
 import pipe.work.Work;
 import routing.Pipe;
 import storage.Storage;
+
+import java.util.ArrayList;
 
 /**
  * Created by rushil on 4/3/16.
@@ -13,19 +18,30 @@ public class Query extends Resource {
 
     Storage.Query query;
 
+    public Query(ChannelQueue sq){
+        super(sq);
+    }
+
+
     public void handleGlobalCommand(Global.GlobalCommandMessage msg) {
 
         query = msg.getQuery();
-        switch (query.getAction()){
-            case GET:
-                break;
-            case STORE:
-                break;
-            case UPDATE:
-            case DELETE:
-                break;
-        }
+        if (msg.getHeader().getDestination() == ((PerChannelGlobalCommandQueue) sq).getRoutingConf().getNodeId()) {
+            switch (query.getAction()) {
+                case GET:
+                    ArrayList<MongoDataModel> arrRespData = MongoDAO.getData("temp",new MongoDataModel(query.getKey(),null));
+                    for(int i=0;i<arrRespData.size();i++){
 
+                    }
+                    break;
+                case STORE:
+                    break;
+                case UPDATE:
+                case DELETE:
+                    break;
+            }
+
+        }
     }
 
     public void handleCommand(Pipe.CommandRequest msg) {
